Don't break gerrit changeset links on linebreak
Using \s instead of space on Regexp of gerrit changeset RegExp to match
on newline, because "gerrit change <number>" is a long string with many
probabilities of getting auto word-wrapped on bugzilla.

Other Regexp may benefit from this like wikilinks (see bug 39237) but I'm
not sure it would cause problems if there's a opening and closing link
intentionally in several lines.

Change-Id: I43aadb0931addd0f12b678160bcc0bf282a4bd44