Remove intro hook <div> from global footer
I have no idea why this is here, but an "intro" hook (which we
don't use) at the bottom of the page feels wrong (should be rather
for the "outro" hook) plus breaks CSS when expanding the page-index
div (tested on boogs.wmflabs.org) with more content.
The "intro" hook still exists at the top of the frontpage (where
it should be).

Change-Id: I0e9827db64378246147f7d358495d7f8fc8e4ee2