Work around Bugzilla XML RPC bug with special Unicode characters
When exporting Bugzilla tickets via Chase's script we run into an API bug with
specific Unicode letters for
https://bugzilla.wikimedia.org/show_bug.cgi?id=9444#c0.
This is applying a hackish upstream workaround described in
https://bugzilla.mozilla.org/show_bug.cgi?id=839023#c10 which is slightly
overkill (applied to the entire output, not just the values provided back when
querying the API). However, that Unicode character range affected by this patch
should not be in use in any relevant way, plus this change affects only XML RPC
behavior (while e.g. Bingle uses JSON RPC, hence not affected).

This file is a copy of the upstream file which has not seen changes since
February 2014 (hence same as in the 4.4.5 tarball code which we have deployed),
with the custom hack being two lines added (search for "69747").

As expressed, not feeling very happy about its scope and cannot 100% judge
potential side effects (but I miss better Perl skills).

Bug: 69747
Change-Id: I960b2c80308213f324968eb44890782488abd4d5