Reapply Bugzilla XML-RPC API workaround for a short while again
Chase asked for this to test mass-fetching numerous Bugzilla tickets
as a preparation for the migration to Phabricator:
We'd like to make sure that there are no further issues we need to
tackle (private comments and attachments, access restricted bug
reports, and non-existing ticket numbers are already handled).
We likely want to remove this again around Wednesday to do further
testing with fetching attachments, then reapply on Friday for
migrating bug reports, then remove for migrating attachments.
Hell yeah.

T815

Change-Id: Id999039c6d0c372235249859f781ae78c6b07e7a